requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dlib==19.19.0
dnspython==1.16.0
email-validator==1.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
protobuf==3.12.2
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
